translation fixes + explicitly defining sectional link id attribute
This describes a fairly widespread issue in the Chinese docs.

The Table of Contents (TOC) extension for mkdocs that generates sectional links within a page does not work well with unicode characters and condenses all of them into a dash, so in the case of the translated Chinese pages, the plug-in was generating sectional links that look like #_1, #_2, which are also causing the internal links to break because a lot of them are still using the old English doc's header id's.

This page on live has quite a few examples of the issue above: https://docs.moonbeam.network/cn/getting-started/local-node/setting-up-a-node/

Current fix is to use an annotation from the python attribute list extension (https://python-markdown.github.io/extensions/attr_list/) to define the sectional header ID's in Chinese language pages so they can be consistent with document sectional links, and also readable.

Will explore in the future to see if there are better solutions, such as using a custom slugify function.